requirements.txt
# AIMC 项目核心依赖
# 此文件包含训练 MineDojo AI 所需的核心依赖
#
# 安装方法:
#   pip install -r requirements.txt
#
# 如需精确复现环境，请使用:
#   pip install -r requirements-freeze.txt

# ==============================================================================
# 核心框架 - 精确版本（必需）
# ==============================================================================
minedojo==0.1               # MineDojo 环境
gym==0.21.0                 # MineDojo/MineCLIP 要求
numpy>=1.21.0,<2.0          # MineDojo 不支持 2.0+
opencv-python==4.8.1.78     # 关键版本！MineRL 窗口显示需要

# ==============================================================================
# 深度学习框架
# ==============================================================================
torch>=2.0.0                # PyTorch（根据系统选择 CPU/CUDA/MPS 版本）
stable-baselines3>=1.8.0    # DAgger 训练框架
tensorboard>=2.9.0          # 训练可视化

# ==============================================================================
# 图像处理
# ==============================================================================
Pillow>=9.0.0               # PIL 图像处理
imageio>=2.9.0              # 视频保存

# ==============================================================================
# 录制工具
# ==============================================================================
pygame>=2.1.0               # 鼠标控制录制（推荐）
pynput>=1.7.0               # 键盘控制录制

# ==============================================================================
# MineCLIP 相关
# ==============================================================================
transformers>=4.20.0        # CLIP Tokenizer
mineclip                    # MineCLIP 视觉-语言模型

# ==============================================================================
# 可视化和进度
# ==============================================================================
matplotlib>=3.5.0           # 数据可视化
tqdm>=4.60.0                # 进度条

# ==============================================================================
# 配置和工具
# ==============================================================================
PyYAML>=6.0                 # YAML 配置文件
psutil>=5.9.0               # 系统监控

# ==============================================================================
# Web 控制台（可选）
# ==============================================================================
Flask>=3.0.0                # Web 服务器
flask-cors>=4.0.0           # CORS 支持

# ==============================================================================
# MineRL 环境（可选）
# ==============================================================================
# minerl==1.0.0              # BASALT 任务
                             # 注意: 需要从 GitHub 安装，详见 docs/guides/MINERL_GUIDE.md

# ==============================================================================
# 其他依赖（会自动安装）
# ==============================================================================
# 以下包是上述包的依赖，会自动安装，无需手动指定:
# - gym3, gymnasium, cloudpickle, dill
# - huggingface-hub, tokenizers, safetensors
# - kornia, einops
# - click, colorama, rich
# - ipython, jedi

# ==============================================================================
# 平台特定说明
# ==============================================================================
# macOS M 系列芯片:
#   - 使用 x86_64 模式: arch -x86_64 /bin/zsh
#   - PyTorch MPS 加速: 自动检测
#   - 参考: DEPLOYMENT.md - 方案B
#
# Linux:
#   - CUDA 版本: 安装对应的 torch
#   - 参考: DEPLOYMENT.md - 方案A
#
# Windows:
#   - 不直接支持，建议使用 WSL2
